Printing the ball
IM is now being able to provide the ball information to the VGA. The
ball coordinates are being stored in the first 20 bits of its LOC
register. First 10 bits for X, second 10 bits for Y. The remaining 12
are unused for now, but can store additional info in the future.

The ball can't be controlled yet though (later, with switches for
testing perhaps).

Also the ball position is not being stored in the actual register - yet.
But it will be.