IOB-TIMER and controllers code added
The timer submodule was added to add the sleep() functionality. Useful
to slow down the program (else it goes as fast as possible).

Also added code to help with the controllers when they're added. Note:
the code may or may not be wrong, depending on how the controllers are
implemented (needs to be checked and tested).