Prepare code to use the new NESCTRL submodule
The NESCTRL submodule is a submodule that contains the NES controller
physical interface. As of this commit, still needs testing. But the C
code should be ready to go.

Also removed test code for the controllers from the IM submodule.

And added warnings and errors flags to the Makefile for GCC.