Ball physics added
The ball now bounces off the paddles and the walls. Currently it also
bounces off the side walls, but this will be changed in the future.

The ball moves always at the same speed, but its angle changes depending
on where it hits the paddles (-45, 0, 45 degrees).

IM submodule implementation improved.

NESCTRL ports corrected.