Various changes
Renamed the IM submodule to PMEM (pseudo-memory), finally a correct
name.

The _LOC registers now have the RGB 12-bit color added to their format,
to complete the 32 bits with information useful to the hardware. Now on
the C struct there are 64 bits for the _LOC registers, but the last 32
are unused by the hardware and are only used by the software.

The ball remains white, while the paddles are now red and green.

Various macros were created in C to ease the programming and be less
confusing.

The velocity is now set in pixels per frame, instead of 0 or 1 for
normal or fast.

The game speed was decreased - the loop is now iterating each 5 ms.

All objects are now moved every time the loop iterates, and their speed
is now taken into account. setCoords() is now internal to GameUtils.c
and to move the paddles, all there is to do is set their speed and
moveObjs() will take care of the rest.

The controller is also now checked for presence by checking if all
buttons are at 1. If they are, then the controller is not present.

Physics corrected (in some way... - I don't remember the difference
anymore...).